+++ age/CHANGELOG.md
@@ -13,6 +13,7 @@ to 1.0.0 are beta releases.
 - `age::encrypted::EncryptedIdentity`
 - Support for the new native age recipient types:
   - `age::tag::Recipient` (encryption-only)
+  - `age::tagpq::Recipient` (encryption-only)
 
 ### Changed
 - MSRV is now 1.74.0.